### 2018-08-09 #888 @whedon
[PRE REVIEW]: cottoncandy: scientific python package for easy cloud storage
**Submitting author:** @anwarnunez (<a href="http://orcid.org/0000-0003-1346-670X">Anwar Nunez-Elizalde</a>)
**Repository:** <a href="https://github.com/gallantlab/cottoncandy" target ="_blank">https://github.com/gallantlab/cottoncandy</a>
**Version:** v.0.0.2
