### 2018-08-09 #888 @anwarnunez
On Thu, Aug 9, 2018 at 12:44 PM, Anwar Nunez-Elizalde <
anwarnunez@berkeley.edu> wrote:

>
> potential reviewers:
>
> Wilfred Tyler Gee: wtgee (wilfred.gee@students.mq.edu.au)
> Vanessa Sochat: vsoch (vsochat@stanford.edu)
> Ariel Rokem: arokem (arokem@gmail.com)
>
>
> On Thu, Aug 9, 2018 at 9:40 AM, whedon <notifications@github.com> wrote:
>
>> --> Check article proof 📄 <--
>> <https://github.com/openjournals/joss-papers/blob/joss.00888/joss.00888/10.21105.joss.00888.pdf>
>>
>> —
>> You are receiving this because you were mentioned.
>> Reply to this email directly, view it on GitHub
>> <https://github.com/openjournals/joss-reviews/issues/888#issuecomment-411821149>,
>> or mute the thread
>> <https://github.com/notifications/unsubscribe-auth/AAU0i3PSx-mP3DnLGEbXgpFK1rjR7LGtks5uPGXngaJpZM4V2Cbi>
>> .
>>
>
>
