### 2018-08-09 #888 @vsoch
A module named cottoncandy? Please let me review this! :candy: :chocolate_bar: :gift:
