### 2018-08-10 281f866
Update paper.md
