### 2018-08-10 #890 @whedon
[REVIEW]: cottoncandy: scientific python package for easy cloud storage
**Submitting author:** @anwarnunez (<a href="http://orcid.org/0000-0003-1346-670X">Anwar Nunez-Elizalde</a>)
**Repository:** <a href="https://github.com/gallantlab/cottoncandy" target ="_blank">https://github.com/gallantlab/cottoncandy</a>
**Version:** 0.1.0

- [x] **Version:** Does the release version given match the GitHub release (v.0.0.2)?
