### 2018-08-11 f1669ad
tests: travis fix
