+++ .travis.yml
@@ -9,7 +9,7 @@ install:
 script: pytest -v
 env:
   global:
-  - secure: UyJENFv2HYtRZU/LPmhXCbrvGT+I9D0iG/YVJGoe0hMEoMLm4/amceH7TaoTVx7o81WRiFnUYV4IFv4jSoeohApSPMeMXZZqJf+S86Tr2g98b4uuglW/Ir0eA3fzmhHPrvXe3SWEAJvTy/JZDxOKRJW79ZEGDlL1btq2ZPmR0rgUF71QbfCwVmon3IoB/1386pBYE09BJpUA6URZAFHJyjEksi9ECQJVkMvw4MEosnk3o08+0HTWRXZA0AWYZOe+v5ZJQFImQs9ctggpAnt0CPhePvG9y+CP/Q3egJZwE2UL8t8r9jz1gt5yl+NXKeC6mZklXa4sCb967r8IwVPAUyA5p4P3MhYY8W/Nhk3RWWvufVJOYLGGGweuisw+Gq2CYXifLQxFW8+UkLLMN8R/7/l28ftDE7s0d8seUh0SFx/kn6KLA3LF3H69zfCxecNRkbwmCMf1f8ep7o71+0V1+NizQSSRYADn4LzuZOwerYmJbOznXC2/Q1x+zreeYocpjoB05ZewV0a848WnPfRPoHA/38SvbtUJU0U5v5W3WqgTJYRvTW+ktMUNialQJZSf2bVWKlnumBb+wUUbYhvbTRTtWqNHM0W9Ly+1vrLNa5BURPQFJst0jrUd4F4m5HdpTgtiPjLDj5DBJmmAwW0k8sqTZYG1DbDj2LJBa1lcDNY=
-  - secure: d6efpPzs6RhjBj8gZ67ez5Rp9CrYFRRixZ6wTTSW9ZitpT5bs1PUJITspByH+GJBImd6y6JZtZuh6OsvJhWcUJjAKrKwEfT791Y5LpT5DIn4CoHUVTBUQlxb6rVJEuy5H82XOepdpUC1cw9keB21IJOojJM5e8n849mMli9Cia7jbdv3exDpk8y+E3WYnKzyUy276h+IH1jTlPDuxIes4NRY4sO8yYS5uDXBNUHET30yCu5sqix1S1izI5F0wNSfrt7JQbJ/EJpB6QuXyUNUwq0uA30J5D4aRjf/5Vsccu95Acw0LzR/t9EZcusc30MgOyOElZpWdA36TgPKNLmxSZhgKQao0crXrMR9eEMluxKuKlA1C9H+xfRglGVOhtBI8KaxmgszpaEQt4XO+tN1bsJ9KOOorisC9rgxxemLB8IwrR1ANozLJnKK4RRyagkwSMAmSqH3OltzsA3pozZ2bc5rUh+SybfaFbFL0cGVXqlmN9bOkc7Xv9gpdZejc6ih34GlHjnHrqZY9bgJqk++XcgvUowFAAV/g7H1AnGz3dOkSfDA2wCU8Zdq4pxFh9TreYyPCGuyYtKoK1H5KcU9Cp6QwDghdqMwJNWwWJG4R97Fklv1nV7qWHqirtnOkY3ahkiuRoeoOJSalKR6AgMOg1u5WEahnSZbs3SquyvfvTE=
[…]
