### 2018-08-11 #890 @vsoch
Thanks @whedon and @leeper ! I'm much looking forward to reviewing this. To give you a time frame, I should be able to get this underway mid to end of this week, finishing up some things this weekend! :)
