### 2018-08-15 3ecf517
logo!
closes #6
