### 2018-08-15 gallantlab/cottoncandy#63 @vsoch
[REVIEW] cottoncandy for Journal of Open Source software
Here I will write my review for https://github.com/openjournals/joss-reviews/issues/890. While the suggestion is for multiple issues, I will try addressing my issues here.

### Repository
As a user, my first interaction with the repository is at https://github.com/gallantlab/cottoncandy

 - I am happy to see the source code, and the instructions for installation front and center. The logo is quite fantastic :)
 - The LICENSE is also present
 - The version string is found in the [setup.py](https://github.com/gallantlab/cottoncandy/blob/master/setup.py#L46).

**change requested** :stars:

For your versioning, please use [semvar](https://semver.org/) convention. Instead of 0.01 for example, it would be 0.0.1 or 0.1.0 (format is `MAJOR.MINOR.PATCH`). Note that this should correspond with the [release here](https://github.com/gallantlab/cottoncandy/releases) (and no v, alas!)

**check requested** :heavy_check_mark:

For authors, I only see 4 on the paper, but I see 11 [contributors here](https://github.com/gallantlab/cottoncandy/graphs/contributors). The authors list likely corresponds to the list here, but could you please clarify who is who, and the contribution (we want to make sure someone with significant contribution is not forgotten!)

### Manuscript

 > is a python scientific library for storing and accessing numpy array data from cloud-based object stores.

Based on the many choices available, I would have expected more than AWS and then `EXPERIMENTAL` Google Drive. I think this should be made more true to that?

 > cottoncandy is a bucket full of syntactic sugar that facilitates the use of cloud storage in typical data science workflows.

 - `S3 API-enabled CEPH` is included in the list, but it's still AWS correct? There is just then AWS and Google Drive (experimental?)

This line is fantastic :)

### User Interaction

#### Installation

As a user, I first go to the repository and want to know how to install the software. Even before reading the instructions I would expect
(given python) to do:

```bash
pip install cottoncandy
```

and the library doesn't seem to be provided. I see that the instructions are to install from Github. A few suggestions / questions!

**question / suggestion** :question:

Do the authors intended to provide the package on pypi, and if not, why? Either way, could the authors also offer an install example
using pip from Github?

I proceeded to (start to) install following instructions to clone:

**change requested**

```
$ git clone https://github.com/gallantlab/cottoncandy.git
$ cd cottoncandy
$ sudo python setup.py install
```

It's generally not good practice to be installing to system python with sudo, or at least it's something I did when new to Python (and eventually never again!) Could the authors please explain the need for system python over:

```bash
# assuming a local or virtual environment with python
python setup.py install

# or assuming a system python and install as a user (common on shared resources)
python setup.py install --user
```

For my review, I have an anaconda environment used as default, so I am able to do `python setup.py install`

**detail requested** :magnify:

Could you explain the logic of your setup.py, using the defaults.cfg and configparser? I am mostly not familiar with this
approach - it looks different than including data with a MANIFEST.in or other? Note the warning:

```bash
$ python setup.py install
/home/vanessa/anaconda3/lib/python3.6/distutils/dist.py:261: UserWarning: Unknown distribution option: 'include_package_data'
  warnings.warn(msg)
```

When you mention:

 > Upon installation cottoncandy will try to find your AWS keys and store them in this file

I'm guessing that the install tries to sniff these from the environment. It might be good to add this detail so the user isn't alarmed about scripts sneaking around in places they should not be. :)

**bug with file generation?**

Note that I would expect from this statement that after install, to find a file at `~/.config/cottoncandy/options.cfg`. I don't see a file (TAB does not complete!) Should there be one?

```bash
cat ~/.config/cottoncandy/options.cfg
```

#### Permissions

Depending on how the user installed it (sudo vs. not) if this file is to be generated, it could be problematic for the effective user home. For example:

```bash
# /root is sudo's home
$ sudo python -c "import pwd; import os; userhome = pwd.getpwuid(os.getuid())[5]; print(userhome)"
/root

# but not my home!
$ python -c "import pwd; import os; userhome = pwd.getpwuid(os.getuid())[5]; print(userhome)"
/home/vanessa
```

> Upon installation cottoncandy will try to find your AWS keys from the environment and store them in this file. If you do not have them exported, you will be able to define them at a later step.

It's also generally nice to know the extent to which a secret is being saved. For example, it's important to me to know if I don't have them exported (but define them later) if they are written to that config file. I may not want this to be the case. The next line then suggests that
I have another option (instead):

> Setup the connection (endpoint, access and secret keys can be specified in the configuration file instead)::

**change request** :delta:

Could you please clarify when the file is generated, where it is generated, where the configuration values come from, and when for the user?

#### Usage

**AWS**

Let's store some stuffs! I'd like to ask if someone from your team could share a testing AWS IAM credential for me to use? I don't
have or use AWS much, and I would need to personally pay for it. I can assure you it will only be used for this test, and if
you want, I think that you can [limit the usage and scope](https://aws.amazon.com/iam/faqs/#iam_role_management_anchor).

When I am able to test the storage, then I can finish the review! Are there other backends (not requiring payment) that I
might test?

**Others?**

I was expecting based on the title to have storage clients beyond s3. I'm also worried about the big `EXPERIMENTAL` (hence why I'm not testing this) in caps, could you give detail as to why these are experimental? It would be good to have them as production ready on the publish of this paper. It's also slightly alarming to see things like:

```
    ##############################
    # USE THIS AT YOUR OWN RISK.
    ##############################
```

### Documentatione

Questions for you!

 - The [version here](https://gallantlab.github.io/cottoncandy/) is also specified differently than the tagged release (there is a v) and not using semvar. Would the authors be able to be consistent between the places?
 - Based on [the upload functions](https://github.com/gallantlab/cottoncandy/blob/99b292ce43bdca0d0440d10a456de251b023815c/cottoncandy/interfaces.py#L520) it looks like you are converting to bytes (with String.io) and then uploading the blog. Then to download, you are reversing the process.
   - How is this going to work between versions of numpy?
   - What is the benefit of doing this for the user, versus allowing them to save (with joblib, pickle, other) their own object to upload?
   - what about different versions and formats of numpy arrays (e.g., pandas or other?)
   - What about versions of Python? (Python 2 vs. Python 3?)

### Api Documentation

This isn't essential, but I do with the documentation pages were as beautifully done as the README, in terms of design, color, etc. A user that is introduced to the software through the portal should get a professional feel.  Otherwise, the functions are here and readthedocs is good!

#### Statement of Need

I definitely there is need for a unified interface to interact with storage, and from python, but I would like to see a little more development
done to provide (minimally) production clients for AWS, Google Drive, and perhaps one more largely used storage backend (my first thoughts
were Google Storage and Azure as well). The statement of need in the manuscript is compelling, but the software itself isn't quite ready
to match it. But it's okay! This can be developed to get cottoncandy in tip top shape.

I think this is a really cool tool for a lab (and the name is fantastic!) but in that it requires wrapping other APIs to interact with services, but the current implementation doesn't provide additional functionality other than dumping and loading a numpy array. For a user that is using AWS or Google storage endpoints, in that these Python libraries (boto, google-storage, google-drive-python etc.) change remarkably quickly, I would advise them to use the libraries directly to save the files as they are generated and avoid the additional layer. The "competitor" for this would be something like [rclone](https://github.com/ncw/rclone). Did the authors consider providing a python wrapper to that?

### Community Guidelines

**Missing guidelines**

Could the authors please add:

 - a CONTRIBUTING.md and CONTRIBUTORS.md, if appropriate
 - a reference / some discussion of how to contribute in the README
 - instructions for how to report an issue in the README
 - (optional) issue and PR templates, if the above is complicated.
