### 2018-08-15 #890 @vsoch
I have [posted my review](https://github.com/gallantlab/cottoncandy/issues/63) and updated the boxes above thus far! Please see the linked issue for discussion on the submission.
