### 2018-08-15 #890 @anwarnunez
Thanks for the review! I'll start working on the fixes address your comments this weekend.
