### 2018-08-15 2dc5ba8
WIP: pip and release versioning
