### 2018-08-15 5870593
DOC: encryption warning
removed experimental flag. added a warning instead. people should backup keys.
