+++ README.md
@@ -115,9 +115,12 @@ See the [Google Drive setup instructions](https://github.com/gallantlab/cottonca
 >>> cci = cc.get_interface(backend='gdrive')
 ```
 
-### Encryption (highly experimental)
+### Encryption
 
-`cottoncandy`provides a transparent encryption interface for AWS S3 and Google Drive. This requires the `pycrypto` package. This is HIGHLY EXPERIMENTAL.
+`cottoncandy`provides a transparent encryption interface for AWS S3 and Google Drive. This requires the `pycrypto` package. 
+
+**WARNING**: Encryption is an advance feature. Make sure to create a backup of the encryption keys  (stored in `~/.config/cottoncandy/options.cfg`). If you lose your encryption keys you will not be able to recover your data!
+
 
 ```python
 >>> import cottoncandy as cc
