### 2018-08-15 fa155c0
TEST: Change to warnings about $$$
