+++ cottoncandy/tests/test_roundtrip_big.py
@@ -37,10 +37,10 @@ if 1:
                            verbose=False)
 else:
     ##############################
-    # USE THIS AT YOUR OWN RISK.
+    # Warning
     ##############################
-    # It will use your defaults to run the tests on.
-    # If you use AWS, you might incurr costs.
+    # This will use your defaults to run the tests on.
+    # If you use AWS, you might incur costs.
     cci = cc.get_interface()
 
 
