### 2018-08-15 #890 @leeper
Feel free to continue the dialogue here or over there as needed and ping me if I can help on any procedural issues.
