+++ requirements.txt
@@ -6,3 +6,5 @@ cloudpickle>=0.2.2
 toolz>=0.7.4
 pycrypto>=2.6.1
 pydrive>=1.3.1
+six>=1.11.0
+dateutil>=2.7.3
