### 2018-08-15 6d766df
DOC: configuration file location
tested on MAC in conda environment
