### 2018-08-15 a53e1d0
DOC: information on how to contribute
