### 2018-08-15 gallantlab/cottoncandy#63 @anwarnunez
Thanks for the comments!
I created a new branch: https://github.com/gallantlab/cottoncandy/tree/joss_review
New release candidate tag: https://zenodo.org/record/1345777
The release candidate is now on pip, too.

I'll address all your other comments in detail this weekend.
Thanks!
Anwar
