+++ requirements.txt
@@ -7,4 +7,4 @@ toolz>=0.7.4
 pycrypto>=2.6.1
 pydrive>=1.3.1
 six>=1.11.0
-dateutil>=2.7.3
+python-dateutil>=2.7.3
