### 2018-08-17 bb8adca
updated compression defaults
