+++ cottoncandy/defaults.cfg
@@ -39,3 +39,10 @@ credentials = credentials.txt
 [encryption]
 method = AES
 key = False
+
+[compression]
+do_compression = True
+# < 2 GB arrays
+small_array = 'gzip'
+# >= 2 GB arrays
+large_array = 'Zstd'
