### 2018-08-17 e6db955
polishing defaults for compression
