### 2018-08-17 ba0de21
fixed quotes present (shouldn't be) in defaults.cfg file
