+++ cottoncandy/defaults.cfg
@@ -43,6 +43,6 @@ key = False
 [compression]
 do_compression = True
 # < 2 GB arrays
-small_array = 'gzip'
+small_array = gzip
 # >= 2 GB arrays
-large_array = 'Zstd'
+large_array = Zstd
