+++ cottoncandy/defaults.cfg
@@ -1,4 +1,5 @@
 [login]
+# set to False if you don't want cottoncandy to find your keys upon installation
 access_key = auto
 secret_key = auto
 endpoint_url = https://s3.amazonaws.com/
