+++ setup.py
@@ -56,7 +56,6 @@ def main(**kwargs):
                   ],
               },
           cmdclass=dict(install=my_install),
-          include_package_data=True,
           long_description = long_description,
           **kwargs)
 
