### 2018-08-20 gallantlab/cottoncandy#63 @anwarnunez
# Response to Review
(https://github.com/openjournals/joss-reviews/issues/890)

Thanks for the detailed review! I've made all the requested changes and provided answers to the questions raised (see branch `joss_review`; https://github.com/gallantlab/cottoncandy/tree/joss_review). The requested changes have definitely made the software better!

I'd like to highlight one point. Our main contribution with cottoncandy is to provide convenience functions to use cloud-storage from within a pythonic environment (e.g. Jupyter) without saving or reading data to disk. This functionality is not provided by libraries recommended (e.g. rcloud, joblib) as far as we are aware. cottoncandy allows the user to seamlessly download and upload to a cloud-based storage solution from within a Jupyter session -- without accessing the user disk.

While cottoncandy currently implements only two back-ends additional back-ends can be added. The AWS S3 API is by far the most widely used cloud-storage solution and can be used outside of AWS (i.e. CEPH). We hope to incorporate support for a Google Cloud Storage back-end in the near future.

In all, we hope to have addressed all the issues raised in the review.

Best,
Anwar

## Responses by point

### Respository

> For your versioning, please use semvar convention. Instead of 0.01 for example, it would be 0.0.1 or 0.1.0 (format is MAJOR.MINOR.PATCH). Note that this should correspond with the release here (and no v, alas!)

**Fixed.** I made a release candidate (0.1.0rc; https://github.com/gallantlab/cottoncandy/releases/tag/0.1.0rc). Once the review is approved, I will push the 0.1.0 release. I could also make a 0.1.0 release now and then a 0.1.1 if there are further changes needed. Let me know what you think.

> For authors, I only see 4 on the paper, but I see 11 contributors here. The authors list likely corresponds to the list here, but could you please clarify who is who, and the contribution (we want to make sure someone with significant contribution is not forgotten!)

**Clarification.** The main developers are myself (AN) and JSG. The following list details the specific significant contributions made by the authors.

* AN: Wrote and designed cottoncandy.
* JSG: Helped design cottoncandy.
* TZ: Implemented Google Drive back-end.
* JLG: Funding.

It is generally difficult to draw sharp authorship lines with software. With respect to the publication of cottoncandy, I believe the list of co-authors reflects all the users who have made significant contributions. The rest of the contributions have been very beneficial(!) but more limited in scope.

### Manuscript

> Based on the many choices available, I would have expected more than AWS and then EXPERIMENTAL Google Drive.

The Google Drive back-end support is not experimental. Apologies for the confusion. It was merged to master over a year ago (Aug 2, 2017). The note on the README was put there when it was first merged. It is now removed.

> I think this should be made more true to that?

This is a good point. The first sentence is a high-level description of the software. Details about the specific cloud-based object stores that are currently supported are listed in the "Long description" section. We wish to include support for Azure and Google Cloud Storage in the near future. I'd like to maintain the generality of the first sentence to reflect the fact that cottoncandy is not specific to any one storage solution. Ideally, once we implement the Google Cloud Storage (or Azure) back-end this point will not be contentious (https://github.com/gallantlab/cottoncandy/issues/64).

> S3 API-enabled CEPH is included in the list, but it's still AWS correct? There is just then AWS and Google Drive (experimental?)

To clarify, CEPH is not AWS. CEPH is an open source storage solution with many capabilities (https://ceph.com/). CEPH can be configured to provide access to stored data using an S3 API (http://docs.ceph.com/docs/mimic/radosgw/). This is what is meant by "S3 API-enabled CEPH".
However, the point raised is well-taken: Only S3 and Google Drive back-ends are provided.

### User interaction

> Do the authors intended to provide the package on pypi, and if not, why? Either way, could the authors also offer an install example

**Fixed.** Sure thing! I've pushed cottoncandy to pip (https://pypi.org/project/cottoncandy/). 0.1.0rc is currently there and I can submit a 0.1.0 once all the changes pertaining to the review are approved. Installation instructions now contain pip as well as the previous method.

> It's generally not good practice to be installing to system python with sudo, or at least it's something I did when new to Python (and eventually never again!) Could the authors please explain the need for system python over:

**Clarification.** There are situations where it is appropriate to install packages system wide (with sudo or root). When managing shared computing resources, it can be impractical to allow multiple virtual environments per user since this will require a lot of disk space. There are also cases where it can be useful to use a bleeding edge package (e.g. master) in production across multiple machine (with shared storage), which can be achieved with `sudo python setup.py develop`.

However, I agree that these use-cases are not widely adopted. I have therefore also provided the more conventional `pip` instructions as suggested.

> Could you explain the logic of your setup.py, using the defaults.cfg and configparser? I am mostly not familiar with this

**Clarification.** The configuration file is much more flexible than a MANIFEST.ni. The configuration file allows the user to specify the default values for some variables in the software. The configuration file is loaded every time the software is imported, not only during installation.
For example, the user can specify the threshold at which multi-part uploads are used (100MB by default; https://github.com/gallantlab/cottoncandy/blob/master/cottoncandy/defaults.cfg#L18), or the default permissions to use for uploaded objects (https://github.com/gallantlab/cottoncandy/blob/master/cottoncandy/defaults.cfg#L9).

> I'm guessing that the install tries to sniff these from the environment. It might be good to add this detail so the user isn't alarmed about scripts sneaking around in places they should not be. :)

**Fixed.** Yes, good point! I have added a more visible warning statement on the README. I can also set this to False by default if the risk/convenience trade-off of this default setting is deemed to high.

>  I don't see a file (TAB does not complete!) Should there be one?

**Details.** Is the software being installed using MAC OS? If so, the configuration file will be stored under `~/Library/Application Support/cottoncandy/options.cfg`. I've added more details to the README (see "Configuration file" subsection). We also took this opportunity to make sure that cottoncandy works on an anaconda environment under MAC OS.

> Depending on how the user installed it (sudo vs. not) if this file is to be generated, it could be problematic for the effective user home.

**Clarification.** Good point! I haven't encountered this problem under Ubuntu (i.e. even under sudo installation, the configuration is for the user not root). I'm not sure how other *NIX systems handle this. I imagine this can be an issue for some OSs. At any rate, using `pip` to install cottoncandy fixes this potential issue as you suggested.

> It's also generally nice to know the extent to which a secret is being saved. For example, it's important to me to know if I don't have them exported (but define them later) if they are written to that config file. I may not want this to be the case.

**Fixed.** I've added a note on the README to inform the user that the keys will be stored in a configuration file. The keys can also be written into the configuration file by the user. Furthermore, the keys and bucket can be specified directly in the  function `cottoncandy.get_interface()`. Again, I can change this default to False if the risk/convenience trade-off of this default setting is deemed too high.

> Could you please clarify when the file is generated, where it is generated, where the configuration values come from, and when for the user?

**Clarification.** The configuration file is created upon installation. The default configuration file (https://github.com/gallantlab/cottoncandy/blob/master/cottoncandy/defaults.cfg) is copied to the user directory (see the "Configuration file" subsection on the README). If the user keys are found upon installation, the keys will also be included in the configuration file. The default configuration file contains the default values for a number of options (https://github.com/gallantlab/cottoncandy/blob/master/cottoncandy/defaults.cfg). The configuration file is read every time cottoncandy is used (upon `import cottoncandy`). This assures that the latest user-defined options are used.

> I'd like to ask if someone from your team could share a testing AWS IAM credential for me to use?

This is tricky because our system is closed-off. However, the tests can also be run using the Google Drive back-end. Setting up Google Drive can be a pain, but we'be included instructions to do so (https://github.com/gallantlab/cottoncandy/blob/master/google_drive_setup_instructions.md). Note that AFAIR Google Drive throttles upload/download speeds to 1Mb.
Currently, automatic testing is set up with Travis CI (https://travis-ci.org/gallantlab/cottoncandy/branches).

If the Google Drive or Travis options do not suffice, please let me know and we can think of other solutions.

> I was expecting based on the title to have storage clients beyond s3. I'm also worried about the big EXPERIMENTAL (hence why I'm not testing this) in caps, could you give detail as to why these are experimental? It would be good to have them as production ready on the publish of this paper. It's also slightly alarming to see things like:

**Fixed.** Apologies for that. I wanted to communicate to users that they may incur costs if they ran the tests themselves (in the order of a couple of cents). I changed this to a less alarming warning (https://github.com/gallantlab/cottoncandy/blob/joss_review/cottoncandy/tests/test_roundtrip.py#L37).

### Documentation

> The version here is also specified differently than the tagged release (there is a v) and not using semvar. Would the authors be able to be consistent between the places?

**Fixed.**  Sorry about that! This is now reflected in the `joss_review` branch: https://github.com/gallantlab/cottoncandy/commit/fd2e6e0441c28ca4916ca1ac485986c84bb0233e

> Based on the upload functions it looks like you are converting to bytes (with String.io) and then uploading the blog. Then to download, you are reversing the process.
[i] How is this going to work between versions of numpy?
[ii] What is the benefit of doing this for the user, versus allowing them to save (with joblib, pickle, other) their own object to upload?
[iii] what about different versions and formats of numpy arrays (e.g., pandas or other?)
[iv] What about versions of Python? (Python 2 vs. Python 3?)

* 1. cottoncandy works across multiple versions of numpy. This is because the numpy array format is stable and does not change across numpy versions (https://github.com/numpy/numpy/blob/master/numpy/lib/format.py). Basically, the array data is stored as bytes with some header information. We are currently separating the header information and storing it in the object metadata with the function `cottoncandy.upload_raw_array()`. The function `cottoncandy.upload_npy_array()` uploads the array data to an object with the header in the byte stream.
* 2. The benefit is huge and this is the main reason to use cottoncandy. cottoncandy eliminates the cost of constantly downloading cloud data to disk and then reading from disk into a Python session. This is not a big issue when using cloud storage as a backup or intermittently. In situations where the cloud storage is the only solution available, the overhead of constantly saving and reading from disk to and from Python sessions becomes prohibitively expensive. (re pickle: cottoncandy provides functionality to stream pickle objects (https://github.com/gallantlab/cottoncandy/blob/joss_review/cottoncandy/interfaces.py#L446)).
* 3. There are only two versions of numpy array formats. Version 2.0 is fully compatible with 1.0 and vice versa except for very large structured arrays (https://github.com/numpy/numpy/blob/master/numpy/lib/format.py#L136). Structured arrays are not supported by cottoncandy, so this is not an issue. However, we do hope to incorporate them soon (https://github.com/gallantlab/cottoncandy/issues/24). When that occurs, a simple flag can easily handle this case.
re pandas: We do not currently support pandas.
* 4. cottoncandy is fully python2 and python3 compatible (Travis build: https://travis-ci.org/gallantlab/cottoncandy)

### API Documentation

> This isn't essential, but I do with the documentation pages were as beautifully done as the README, in terms of design, color, etc. A user that is introduced to the software through the portal should get a professional feel. Otherwise, the functions are here and readthedocs is good!

Ah, yes! Apologies. I haven't gotten around to making them prettier and uploading the logo! Hopefully soon =)

#### Statement of Need

> I definitely there is need for a unified interface to interact with storage, and from python, but I would like to see a little more development done to provide (minimally) production clients for AWS, Google Drive, and perhaps one more largely used storage backend (my first thoughts were Google Storage and Azure as well). The statement of need in the manuscript is compelling, but the software itself isn't quite ready to match it. But it's okay! This can be developed to get cottoncandy in tip top shape.

cottoncandy has been used in production in our lab for more than 3 years (on Github for 2). We have stored approximately 181TB of data using cottoncandy. Similarly, the Google Drive has been used in production for the last year for conducting data backups.
But the point is well taken. We do wish to include support for more back-ends. Google Storage actually has very similar semantics to S3 (https://cloud.google.com/storage/docs/boto-plugin#uploading-objects). Supporting that back-end should be straightforward. I have created an issue for this (https://github.com/gallantlab/cottoncandy/issues/64).

> I think this is a really cool tool for a lab (and the name is fantastic!) but in that it requires wrapping other APIs to interact with services, but the current implementation doesn't provide additional functionality other than dumping and loading a numpy array.

**Details.** Just to clarify: cottoncandy doesn't dump/load numpy array data in the disk storage sense. It streams to and from the object store into memory directly. This also applies to other data formats (e.g. pickle, JSON).

> For a user that is using AWS or Google storage endpoints, in that these Python libraries (boto, google-storage, google-drive-python etc.) change remarkably quickly, I would advise them to use the libraries directly to save the files as they are generated and avoid the additional layer.

**Response.** I agree with the problem. There is a lot of overhead in using these libraries that are not immediately apparent. cottoncandy actually provides a solution to these problems.
For example, there is a lot of boiler plate code that needs to be written when uploading large data with boto to AWS. This is handled using a multi-part uploads (MPU). cottoncandy automatically handles the MPU for the user (https://github.com/gallantlab/cottoncandy/blob/master/cottoncandy/s3client.py#L383).  cottoncandy also handles client-side streaming compression and decompression, which decreases storage and bandwidth requirements. This is implemented to require no more memory than the size required for the buffer (https://github.com/gallantlab/cottoncandy/blob/joss_review/cottoncandy/utils.py#L429).

> The "competitor" for this would be something like rclone. Did the authors consider providing a python wrapper to that?

I didn't know about this. Looks like a great package! Alas, from what I can see, rclone depends on disk storage: "Rclone is a command line program to sync files and directories to and from".
This would also involve reading/saving from disk and then to the object store.

### Community Guidelines

**Fixed.** I added a "Contributing" section to the README.
