### 2018-08-20 gallantlab/cottoncandy#63 @anwarnunez
openjournals/joss-reviews#890
