### 2018-08-20 gallantlab/cottoncandy#63 @vsoch
# Reviewer Response

I will start the second part of my review - for reference I have [cloned this branch](https://github.com/gallantlab/cottoncandy/tree/joss_review) and installed cottoncandy as follows:

```
$ git clone -b joss_review https://github.com/gallantlab/cottoncandy
$ cd cottoncandy
$ python setup.py install
```

## Review Metadata

 - Response to Review
 - openjournals/joss-reviews#890
 - [Github](https://www.github.com/gallantlab/cottoncandy/tree/joss_review)

## Reviewer Response

 - I too agree that the software is much better, solid work @anwarnunez!

> I'd like to highlight one point. Our main contribution with cottoncandy is to provide convenience functions to use cloud-storage from within a pythonic environment (e.g. Jupyter) without saving or reading data to disk. This functionality is not provided by libraries recommended (e.g. rcloud, joblib) as far as we are aware. cottoncandy allows the user to seamlessly download and upload to a cloud-based storage solution from within a Jupyter session -- without accessing the user disk.

I am going to push you further on this - to either provide additional "main contribution" points, evidence that cottoncandy does this better than existing options, **or** just mentioning that they exist (and then writing about the other options in your paper and how they compare). I can offer to help, because I've used some of the other Python software, and notably they allow for writing of objects (as blobs) to storage without needing to interact with the local filesystem. For example, if you look in the [Google Cloud Storage](https://cloud.google.com/appengine/docs/standard/python/googlecloudstorageclient/read-write-to-cloud-storage#writing_to_cloud_storage) getting started tutorial, you see use of the library [cloudstorage](https://pypi.org/project/cloudstorage/). This library of course has:

> convenience functions to use cloud storage

```bash
from cloudstorage.drivers.<driver> import *
```

> without reading data to disk

This logically isn't needed, because the operation before uploading a "blob" to storage is already reading in the file. E.g. you read in as binary data, and then stream from the file, directly as a stream, or a string object. If you look at the various library implementations, they are very general (any binary / blob), so maybe the cottoncandy "spin" would be to better develop the functions to be for data science objects? I don't know if this is a good idea (it might be terrible) but if the other libraries do a good job of loading / unloading general blobs from Python (could be dict or numpy array, we don't care) I would ask what they **don't** do well. They don't do well to capture metadata about the data frame, such as the environment it was generated in, versions of things, or maybe even metadata about the table / headers (if pandas). Cottoncandy would add a layer of richness (and novely, utility) if it offered these kinds of things beyond just moving around blobs.

There are notably functionality that (possibly) is provided via the above (or the underlying libraries) that cotton candy doesn't provide. For example, exponential backoff is normally done with the retrying library, and if the other solutions **don't** offer it, cottoncandy might. I am not sure about the current state of this, so I [posted an issue](https://github.com/scottwernervt/cloudstorage/issues/15) so we both can find out. I'm curious - if cottoncandy doesn't have it, how often was it that you had issues with uploading due to network, etc.?

Other than the individual clients, another notable wrapper akin to cotton candy is [Apache libcloud](https://libcloud.apache.org/). If you do a Google search a few similar pop up too. Anyway - I don't mean to say you need to invent a completely new thing, but to properly review the alternatives, the features that are notably different, and (the goal) is to give the rationale that I as a user would choose cottoncandy over the other options. From my general review, I don't see that cottoncandy offers me anything more than what is available, and maybe is limited in not having exponential backoff and a broader arrange of client options.

> While cottoncandy currently implements only two back-ends additional back-ends can be added. The AWS S3 API is by far the most widely used cloud-storage solution and can be used outside of AWS (i.e. CEPH). We hope to incorporate support for a Google Cloud Storage back-end in the near future.

If adding a backend comes down to supporting the underlying modules for it, isn't this what the cloudstorage module is already doing? Could you use this as a base and add the "syntactical sugar" functions to it to provide metrics about the dataframes, etc. for the user?

## Responses By Point

 - **Versioning** Looks fantastic! Thank you for fixing this up. I am in agreement about having the 0.1.0rc and then 0.1.0 when the first release is published (or whichever marginal increment from that winds up being the chosen one :P).
 - **Authors** Thank you for the clarification, and I definitely understand this is a challenging point! I don't know if @whedon and @arfon have a good place for this in mind, but lots of journals have some kind of "who did what" summary, maybe JoSS could have a suggested way to do this too? I had originally been biased to say "Well, the record of contribution is in Github commits!" but on the other hand, the humanly generated list is meaningful to say "These are the contributors we assert to have made significant contribution to warrant being on the paper, and here is why."
 - Thank you for the clarification on CEPH and the configuration files!
 - [https://pypi.org/project/cottoncandy/](https://pypi.org/project/cottoncandy/) great!

> Clarification. There are situations where it is appropriate to install packages system wide (with sudo or root). When managing shared computing resources

In research computing we use a module system where it's (still) not a system python, but one installed custom. It does require root given the location, but arguably the documentation should be written with the user in mind that would be installing either to python on a shared resources (where he/she doesn't have root) or encouraged to be working reproducibly with something like a virtual environment, in which case root **still** wouldn't be used. I can see arguments for both, but I see stronger case for encouraging the user to not use it.

> However, I agree that these use-cases are not widely adopted. I have therefore also provided the more conventional pip instructions as suggested.

Ah, ok sounds good :)

### Installation
Okay, here is some more digging into installation and configuration issues. First an edge case of having installed with python setup.py install, and then trying with pip:

```bash
$ pip install cottoncandy
Requirement already satisfied: cottoncandy in /home/vanessa/anaconda3/lib/python3.6/site-packages (0.1)
```
Note that the install above is with the `joss_review` branch and I cannot find the configuration file in the expected location. and **Clarification** I'm not on MacOS - I'm running from Ubuntu 16.04. I'm using Anaconda, however. Should I look elsewhere? Next I tried
uninstalling (with pip)

```
vanessa@vanessa-ThinkPad-T460s:~/Documents/Dropbox/Research/Reviews/2018/joss/cottoncandy-gallantlab$ pip uninstall cottoncandy
Cannot uninstall 'cottoncandy'. It is a distutils installed project and thus we cannot accurately determine...
```

So I removed manually. And then did a clean install with pip (note that this is pip from anaconda). The configuration file still wasn't there!
So I tried pip with the --user directive:

```bash
pip install --user cottoncandy
```

It's still not there. What am I doing wrong? After I can find this file I can look over this point again:

> It's also generally nice to know the extent to which a secret is being saved. For example, it's important to me to know if I don't have them exported (but define them later) if they are written to that config file. I may not want this to be the case.
> Fixed. I've added a note on the README to inform the user that the keys will be stored in a configuration file. The keys can also be written into the configuration file by the user. Furthermore, the keys and bucket can be specified directly in the function cottoncandy.get_interface(). Again, I can change this default to False if the risk/convenience trade-off of this default setting is deemed too high.

I could test out Drive once we have more discussion on the above. I'd like to hold off final testing until we are sure it's good to go!

Contributing looks good, and my other issues are addressed! Looking forward to hearing more!
