### 2018-08-20 #890 @vsoch
Great job @anwarnunez and thank you @leeper and @whedon. My further comments are here:

https://github.com/gallantlab/cottoncandy/issues/63#issuecomment-414398356

Happy Monday!
