### 2018-08-20 gallantlab/cottoncandy#63 @anwarnunez
Thanks for the comments!
I'll reply in more detail later in the week.
For now, I just wanted to apologize for the conda install problems.
I've made an issue (#65) and will get this sorted ASAP.
