### 2018-08-21 34e5d96
BF: configuration file upon install
