### 2018-08-21 gallantlab/cottoncandy#63 @vsoch
Tada! Update - it is correct that the config is created on first import! I just did a fresh install (no config) and then imported, exited, and then...
```bash
$ ls /home/vanessa/.config/cottoncandy/options.cfg
  /home/vanessa/.config/cottoncandy/options.cfg
```
This probably makes sense, because if an admin is installing globally for users, this would show up (for the individual user) on import. So maybe let's make this stated clearly in the instructions, and also a question - is there a need for an "admin config" sort of deal, and then a user config to customize further?  For example, the admin cares about upload / download / encryption stuffs, but the user (mostly) cares about credentials I would guess. The next question would be priority - should the admin's defaults take preference? This is (sort of) an approach that Singularity takes - the user has total decision to set binds, and similar, but the admin gets to control how that works.
