### 2018-08-21 gallantlab/cottoncandy#63 @vsoch
See https://github.com/scottwernervt/cloudstorage/issues/15#issuecomment-414699805 about exponential backoff - it looks like boto3 might implement it, so let's take note that whatever libraries are used in cottoncandy, the effort is taken to have exponential backoff (and maybe if possible, multiprocessing for downloads?) Also- what about atomicity? I should have mentioned these things earlier, just remembering now :)
