### 2018-08-21 652929e
bug squashing (bad order of if statements)
