### 2018-08-21 gallantlab/cottoncandy#63 @vsoch
> re configuration options for admins: Yes, I can make that happen! We've had this kinda setup before. A system-wide install can have a options.cfg file in the site/dist-packages directory. These values are copied to the user options.cfg upon first import. The user can thereafter modify these values from their ~/.config/cottoncandy/options.cfg. If a value in the user options does not exist, it is read from the system-wide options.cfg (e.g. encryption keys).

Awesome! Let's make sure this is nicely documented so an admin and user know:

 1. what happens with respect to writing files
 2. when it happens
 3. where they are! :)

An admin can specify system-wide values by modifying the defaults.cfg file from the cloned repo and then performing a sudo python setup.py install kind of thing. This is along the lines of how numpy handles custom builds (https://github.com/numpy/numpy/blob/master/site.cfg.example). The difference is that in this setup the user's values get priority, not the admin. In this case, the user can still overwrite defaults =S. How do you implement that with Singularity?

I think in the admin file it's easiest to have booleans that determine to allow (or disallow) user control of a variable or resource. E.g., [see here](https://github.com/singularityware/singularity/blob/master/etc/singularity.conf.in#L15). You can also have filters that limit what they are able to select from, and then when you load up the libraries just parse the admin config first, determine the logic, then the user config to finish up figuring out settings.

> re download issues: Yup, downloads can fail when the network is saturated (or if using AWS, whenever AWS is acting up). boto allows the user to specify specify timeouts and retries (It appears this is implemented with exponential delays: https://github.com/boto/botocore/blob/develop/botocore/data/_retry.json#L87 ). One can already combine these options with cottoncandy:

> import cottoncandy as cc
> from botocore.client import Config
> config = Config(connect_timeout=60, read_timeout=60, retries=dict(max_attempts=10))

> cci = cc.get_interface('mybucket', config=config)

okay, I'm satisfied with this example! Let's make sure it's easy for a user to find if he/she wants to do the same.

> re multipart download: Yes! Sadly, this has been on my to-do list for 2 years =( I just haven't had the bandwidth to implement it (#11).

Maybe there's an open source developer that could contribute to the library after this publication? *wink*

> re atomicity: Not sure I know what this means exactly. Are you referring to checks on when objects are uploaded? At the moment, this is all left to the back-end API. For S3, the object is not updated unless the upload is successful. For Google Drive, I imagine the same but I haven't tested this.

This is related to downloading - making sure two processes don't try and write to the same file object at once. It would go with the previous point, so don't need to address (yet).
