### 2018-08-22 521266e
setup: configuration options and user info
