+++ requirements.txt
@@ -1,3 +1,4 @@
+wheel
 numpy>=1.6.0
 scipy>=0.9.0
 boto3>=1.2.3
