### 2018-08-22 28691ba
NF: headless instantiation
