+++ cottoncandy/defaults.cfg
@@ -1,6 +1,6 @@
 [login]
-access_key = auto
-secret_key = auto
+access_key = False
+secret_key = False
 endpoint_url = https://s3.amazonaws.com/
 verbose_boto = False
 
