### 2018-08-22 7a114b0
DOC: configuration file
