### 2018-08-23 gallantlab/cottoncandy#63 @vsoch
These changes look great! I’m at the doctor now but I would like to give one more look over and test when I get home, and likely sign off after that. Overall really great work on the fixes!
