### 2018-08-23 gallantlab/cottoncandy#63 @vsoch
## config

omg I love this!

``
In [1]: import cottoncandy

############################################################
Hi! Looks like this is your first time using cottoncandy.

Your cottoncandy configuration file will be stored in:
/home/vanessa/.config/cottoncandy/options.cfg

You can store your S3 or Google Drive credentials there,
and many other options.

Thanks for using cottoncandy!
```

## defaults
I think this is good for now, because it's more likely that a single user is going to use it vs. having it installed at the system level (by an admin) and then used across users. If/when the time comes for this, it will be natural for someone to open an issue and you can address then. Given this, the atomicity issue is also not likely an issue, so I'm good :)

I'm going to look over the compiled pdf again, and will continue in the main thread!
