### 2018-08-23 #890 @vsoch
@leeper my review is complete, I think we are good to go! @anwarnunez and @leeper we will need the release to be 0.1.0 to coincide with the rc versions, and the (updated) one in this issue board. Great work everyone!
