### 2018-08-23 65e5c2c
minor: setup.py version change
