### 2018-08-23 #890 @anwarnunez
Great!
All done from my end, too.
I merged, tagged and uploaded to pip:
https://pypi.org/project/cottoncandy/
https://github.com/gallantlab/cottoncandy/releases/tag/0.1.0
