### 2018-08-24 #890 @leeper
@anwarnunez Please merge this PR: https://github.com/gallantlab/cottoncandy/pull/68
